show node modules size before nuking

+++ .zshrc
@@ -98,7 +98,7 @@ wtf-node-modules() {
 }
 
 clean-node-modules() {
-  find . -name 'node_modules' -type d -prune -print -exec rm -rf '{}' \;
+  wtf-node-modules && find . -name 'node_modules' -type d -prune -print -exec rm -rf '{}' \;
 }
 
 clean-terraform() {
